requirements.txt
ddddocr
pycryptodome
requests
